Fixup serialization
