+++ libraries/chain/stored_value.cpp
@@ -119,3 +119,26 @@ namespace graphene { namespace chain {
 
 } } // graphene::chain
 
+namespace fc {
+
+void from_variant( const fc::variant& var, graphene::chain::stored_debt& value, uint32_t max_depth )
+{
+   FC_THROW_EXCEPTION( fc::assert_exception, "Unsupported!" );
+}
+
+void from_variant( const fc::variant& var, graphene::chain::stored_value& value, uint32_t max_depth )
+{
+   FC_THROW_EXCEPTION( fc::assert_exception, "Unsupported!" );
+}
+
+void to_variant( const graphene::chain::stored_debt& value, fc::variant& var, uint32_t max_depth )
+{
+   to_variant( value.get_value(), var, max_depth );
+}
+
+void to_variant( const graphene::chain::stored_value& value, fc::variant& var, uint32_t max_depth )
+{
+   to_variant( value.get_value(), var, max_depth );
+}
+
+} // fc
